+++ src/renderers/TableRenderer.php
@@ -9,7 +9,7 @@
 namespace unclead\widgets\renderers;
 
 use yii\base\InvalidConfigException;
-use yii\db\ActiveRecord;
+use yii\db\ActiveRecordInterface;
 use yii\helpers\ArrayHelper;
 use yii\helpers\Html;
 use unclead\widgets\components\BaseRenderer;
@@ -133,7 +133,7 @@ class TableRenderer extends BaseRenderer
      * Renders the row content.
      *
      * @param int $index
-     * @param ActiveRecord|array $item
+     * @param ActiveRecordInterface|array $item
      * @return mixed
      * @throws InvalidConfigException
      */
@@ -160,8 +160,7 @@ class TableRenderer extends BaseRenderer
             $hiddenInputs = implode("\n", $hiddenInputs);
             $cells[0] = preg_replace('/^(<td[^>]+>)(.*)(<\/td>)$/s', '${1}' . $hiddenInputs . '$2$3', $cells[0]);
         }
-
-
+        
         $content = Html::tag('tr', implode("\n", $cells), $this->prepareRowOptions($index, $item));
 
         if ($index !== null) {
@@ -175,7 +174,7 @@ class TableRenderer extends BaseRenderer
      * Prepares the row options.
      *
      * @param int $index
-     * @param ActiveRecord|array $item
+     * @param ActiveRecordInterface|array $item
      * @return array
      */
     protected function prepareRowOptions($index, $item)